+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 * Improve UI (`#157 <https://github.com/qbicsoftware/sample-tracking-status-overview/issues/157>`_)
 
+* Add new column with checkboxes to subscribe/unsubscribe directly for each project (`#162 <https://github.com/qbicsoftware/sample-tracking-status-overview/issues/162>`_)
+
 **Fixed**
 
 * Address error when double-clicking a project with no samples (`#164 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/164>`_)